refactor - organizando componentes
