+++ front-end/src/components/NavBar/NavBar.jsx
